+++ AddOns/Baganator/Core/Utilities.lua
@@ -52,6 +52,7 @@ local itemFrame = CreateFrame("Frame")
 itemFrame.elapsed = 0
 itemFrame:SetScript("OnEvent", function(_, _, itemID)
   if pendingItems[itemID] ~= nil then
+    addonTable.ReportEntry()
     local forItemID = pendingItems[itemID]
     pendingItems[itemID] = nil
     for _, callback in ipairs(forItemID) do
@@ -81,4 +82,17 @@ function addonTable.Utilities.LoadItemData(itemID, callback)
   itemFrame:RegisterEvent("ITEM_DATA_LOAD_RESULT")
   itemFrame:SetScript("OnUpdate", itemFrame.OnUpdate)
   C_Item.RequestLoadItemDataByID(itemID)
+end
+
+function addonTable.Utilities.ChatInsertLink(link)
+  if link ~= nil then
+    if ChatFrameUtil and ChatFrameUtil.InsertLink then
+      if not C_ChatInfo.InChatMessagingLockdown or not C_ChatInfo.InChatMessagingLockdown() then
+        return ChatFrameUtil.InsertLink(link)
+      end
+    else
+      return ChatEdit_InsertLink(link)
+    end
+  end
+  return false
 end