+++ AddOns/Baganator/ItemViewCommon/BankViewWarbandView.lua
@@ -309,7 +309,7 @@ function BaganatorItemViewCommonBankViewWarbandViewMixin:SetupBlizzardFramesForT
     -- Ensure right-clicking a bag item puts the item into this bank
     (AccountBankPanel or BankPanel).selectedTabID = bagID;
     if Syndicator.Constants.CharacterBankTabsActive then
-      BankFrame.BankPanel.bankType = Enum.BankType.Account
+      BankFrame.BankPanel:SetBankType(Enum.BankType.Account)
     else
       BankFrame.activeTabIndex = addonTable.Constants.BlizzardBankTabConstants.Warband
       BankFrame.selectedTab = 1
@@ -524,7 +524,7 @@ function BaganatorItemViewCommonBankViewWarbandViewMixin:ShowTab(tabIndex, isLiv
     if self.isLive and C_Bank.CanPurchaseBankTab(Enum.BankType.Account) then
       self.BankMissingHint:SetText(addonTable.Locales.WARBAND_BANK_NOT_PURCHASED_HINT)
     elseif self.isLive and self.isLocked then
-      self.BankMissingHint:SetText(ACCOUNT_BANK_LOCKED_PROMPT)
+      self.BankMissingHint:SetText(BANK_LOCKED_REASON_BANK_DISABLED)
     elseif self.isLive then
       self.BankMissingHint:SetText(addonTable.Locales.WARBAND_BANK_TEMPORARILY_DISABLED_HINT)
     else
@@ -544,7 +544,6 @@ function BaganatorItemViewCommonBankViewWarbandViewMixin:ShowTab(tabIndex, isLiv
 
   self:GetParent().AllButtons = {}
   tAppendAll(self:GetParent().AllButtons, self:GetParent().AllFixedButtons)
-  tAppendAll(self:GetParent().AllButtons, self.LiveButtons)
 
   local sideSpacing, topSpacing = addonTable.Utilities.GetSpacing()