+++ AddOns/Baganator/ItemViewCommon/ContainerSlots.lua
@@ -5,7 +5,8 @@ local swapTracker = CreateFrame("Frame")
 
 function ApplyCursor(targetInventorySlot, associatedTargetBag)
   local location = C_Cursor.GetCursorItem()
-  if location == nil or not C_Item.DoesItemExist(location) or select(6, C_Item.GetItemInfoInstant(C_Item.GetItemID(location))) ~= Enum.ItemClass.Container then
+  -- Handle bags dragged from other equipment slots or items dropped into bags
+  if location == nil or not location:HasAnyLocation() or not C_Item.DoesItemExist(location) or select(6, C_Item.GetItemInfoInstant(C_Item.GetItemID(location))) ~= Enum.ItemClass.Container then
     PutItemInBag(targetInventorySlot)
     return
   end
@@ -459,7 +460,6 @@ function BaganatorBagSlotsContainerMixin:OnLoad()
     else
       bb:SetPoint("TOPLEFT", self.liveBagSlots[#self.liveBagSlots - 1], "TOPRIGHT")
     end
-    addonTable.Utilities.MasqueRegistration(bb)
     addonTable.Skins.AddFrame("ItemButton", bb, {"containerBag"})
   end
 
@@ -478,7 +478,6 @@ function BaganatorBagSlotsContainerMixin:OnLoad()
   for index = 1, bagSlotsCount do
     local bb = GetCachedBagSlotButton()
     bb.SlotBackground:SetTexture((select(2, GetInventorySlotInfo("Bag1"))))
-    addonTable.Utilities.MasqueRegistration(bb)
     addonTable.Skins.AddFrame("ItemButton", bb, {"containerBag"})
     bb:UpdateTextures()
     bb.isBag = true